fix/Nav
fixing props
