+++ src/components/Nav/Nav.jsx
@@ -1,3 +1,5 @@
+//@ts-nocheck
+import {useState} from 'react'
 import {NavLink, useLocation} from 'react-router-dom'
 import PropTypes from 'prop-types'
 
@@ -9,7 +11,9 @@ import navigationRoutes from './index'
 import icon from '../../assets/svg/sprite.svg'
 import css from './Nav.module.css'
 
-const Nav = ({setNavIsOpen, setDropIsOpen}) => {
+const Nav = ({setNavIsOpen}) => {
+  const [isDropOpen, setDropIsOpen] = useState(false)
+
   const location = useLocation()
 
   const {routes, landingPageRoutes} = navigationRoutes
@@ -19,6 +23,10 @@ const Nav = ({setNavIsOpen, setDropIsOpen}) => {
   const handleNavLinkClick = () => {
     setNavIsOpen(false)
     setDropIsOpen(false)
+  }
+
+  const handleDropdownToggle = () => {
+    setDropIsOpen(!isDropOpen)
   }
 
   return (
@@ -34,14 +42,36 @@ const Nav = ({setNavIsOpen, setDropIsOpen}) => {
             key={title}
             className={css.navItem}
             to={href}
-            onClick={() => (el ? null : handleNavLinkClick())}
+            onClick={() => {
+              if (el) {
+                handleDropdownToggle()
+              } else {
+                handleNavLinkClick()
+              }
+            }}
           >
             {title}
-            {el && <Dropdown setNavIsOpen={setNavIsOpen} />}
+            {el && (
+              <div className={css.dropdown}>
+                <svg className={css.dropdownIcon}>
+                  <use href={`${icon}#dropdown`}></use>
+                </svg>
+                <ul
+                  className={css.dropdownList}
+                  style={{display: isDropOpen ? 'block' : 'none'}}
+                >
+                  <Dropdown
+                    setNavIsOpen={setNavIsOpen}
+                    setDropIsOpen={setDropIsOpen}
+                  />
+                </ul>
+              </div>
+            )}
           </NavLink>
         ))}
         <Button variant="support" content="Wesprzyj" />
       </nav>
+
       <div className={css.navLang}>
         <button className={css.navLangBtn}>PL</button>
         <span className={css.navLangSpan}>|</span>
@@ -52,8 +82,7 @@ const Nav = ({setNavIsOpen, setDropIsOpen}) => {
 }
 
 Nav.propTypes = {
-  setNavIsOpen: PropTypes.func.isRequired,
-  setDropIsOpen: PropTypes.func.isRequired,
+  setNavIsOpen: PropTypes.func,
 }
 
 export default Nav
