+++ src/components/Nav/Nav.module.css
@@ -73,8 +73,27 @@
 .navLangSpan {
   margin: 0 4px;
 }
+
+.dropdown {
+  display: flex;
+  flex-direction: column;
+}
+
 .navDropdownIcon {
   width: 10px;
+}
+
+.dropdownIcon {
+  margin-top: 11px;
+  width: 11px;
+  height: 11px;
+  margin-left: 10px;
+}
+
+.dropdownList {
+  display: none;
+  margin-top: 20px;
+  margin-left: -60px;
 }
 
 @media screen and (min-width: 765px) {
@@ -157,5 +176,18 @@
     right: 0;
     margin-top: 4px;
     margin-right: 8px;
+  }
+
+  .dropdownIcon {
+    margin-top: 7px;
+  }
+
+  .dropdownList {
+    margin-left: -66px;
+    margin-top: 34px;
+    background-color: var(--primary-background-color);
+    position: fixed;
+    z-index: 999;
+    border-radius: 0px 0px 12px 12px;
   }
 }
